move syntax tests to new file syntax.jl
