+++ test/choosetests.jl
@@ -49,7 +49,7 @@ function choosetests(choices = [])
         "checked", "bitset", "floatfuncs", "compile", "distributed", "inline",
         "boundscheck", "error", "ambiguous", "cartesian", "asmvariant", "osutils",
         "channels", "iostream", "specificity", "codegen", "codevalidation",
-        "reinterpretarray"
+        "reinterpretarray", "syntax"
     ]
 
     if isdir(joinpath(JULIA_HOME, Base.DOCDIR, "examples"))
